+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+* CXV1-8474 - Fixed issue that was allowing the user to copy a flow with no active version and have the flow set to 'enabled'
+
 ## [4.10.13]
 * CXV1-6501 - fixed broken enable/disable functionality, also fixed bug with drafts table not updating on save of new draft
 * CXV1-8450 - Bumped Realtime Dashboards again to add several new widgets on both interactions and resources dashboards