+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * CXV1-8474 - Fixed issue that was allowing the user to copy a flow with no active version and have the flow set to 'enabled'
 * Bumping Flow Designer
 * CXV1-8599 - Bumping soundwave
-* CXV1-2085 - Added white label branding panel.
+* CXV1-2085 - Added white label branding panel and bumping config-shared to match
 
 ## [4.10.13]
 * CXV1-6501 - fixed broken enable/disable functionality, also fixed bug with drafts table not updating on save of new draft