+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Bumping Flow Designer
 * CXV1-8599 - Bumping soundwave
 * CXV1-2085 - Added white label branding panel and bumping config-shared to match
+* CXV1-8082 - Change from sandbox to production instance for salesforce interaction lib
 
 ## [4.10.13]
 * CXV1-6501 - fixed broken enable/disable functionality, also fixed bug with drafts table not updating on save of new draft